Major refactor
Though the script's main functionality is
at this point, broken, the UI loads and
the refactored bits will be reconnected
eventually.

The folder structure was adapted to play nice
with Maya 'modules'. There's a lot of 'nested'
folders, but at least this way the script
can be easily distributed AND doesn't pollute
the global namespace.

Parts of the script, especially the UI,
have been split up into submodules.

A new helper class was added:
`ObjectSpaceBoundingBox`

The above is pretty self-descriptive, and
will be useful for debugging the placement
of the wrapper and object to be wrapped.

The readme was also updated.